Fix builder calls on eloquent builder
We need to also apply the builder mixins to the eloquent builder.
Otherwise it will use the `forwardCall` method and return itself
in the end. But we need to be able to return mixed values instead, eg.
for the toGeojsonFeatureCollection method.

See https://github.com/laravel/framework/issues/21950#issuecomment-437887175

Fixes #11