Remove DB hook to fix having sql in binding values
We currently tried to convert bindings, that were geometries, to an SQL
representation. This does not work, since bindings are passed as
strings to the DB. Thus it cannot execute functions like ST_GeomFromEWKT
in such a binding.

So to fix this issue we will for now remove our custom DB hook and let
PHP bind our geometries as strings. So it will now implicitly use the
string generator.

This method works for all generators and we might consider adding a
custom stUpdate command in the future once we need more explicit
control.