fix(geometries): Fix stringifyFloat to correctly format number
We missed the "." here and thus PHP formatted numbers with the default precision of 6 and prefixed it with whitespaces.
See PHP docs: https://www.php.net/manual/en/function.sprintf.php

This has been addressed by using the `json_encode` function to print the
float. This function uses PHP's `serialize_precision` instead, which
by default uses a similar algorithm to the one PostGIS uses for
printing out the float in a minimal representation and also
allows to print out at the maximum precision, as long as the
user did not change the serialization precision config.

Thanks @ronnypolloqueri for reporting this.