fix(builder): numeric SRID SQL queries
When using numeric values for the SRID in functions like `ST_Transform`, the query will throw an error that it cannot parse the SRID projection.

This seems to be caused by the PDO driver, which internally always converts parameters to strings.
See https://stackoverflow.com/questions/66625661/cannot-bind-value-as-int-with-pdo-pgsql-driver

We now append a `::int` cast to numeric SRID parameters, so they work as expected again.

For Closures and Expressions, the developer needs to take care of this. We added a note to make them aware of this behavior.