feat!: support laravel 12 & php8.4, remove L9, L10 & PHP 8.1 support
This updates our minimum supported Laravel & PHP versions to 11.x and 8.2.
It also adds support for Laravel 12.x and PHP 8.4.

With this we also could remove the DBAL types, as newer Laravel versions do not use DBAL at all and offer no method to register the types.

With this, we also had to update some dependencies like Larastan and Pest. So we fixed all 4 new Larastan issues, mostly minor and one included the non-null byteOrder field on the Scanner class. It concluded that this is never null in usage, so we adjusted that.

The validation rule was adjusted to use the normal ValidationRule contract instead of the deprecated invokable rule.

We also included the `tests/Extra` folder in the phpstan config to allow for an analysis of the traits, which are otherwise unused and ignored by phpstan, as they need a context to be analyzed.